+++ etc/make_release.py
@@ -88,7 +88,7 @@ ISSUE_TYPE_ID = {'Backport': '10300',
               show_default=True,
               help='The remote reference which points to the mongodb/mongo-cxx-driver repo')
 @click.option('--c-driver-build-ref',
-              default='1.30.1',
+              default='2.0.0',
               show_default=True,
               help='When building the C driver, build at this Git reference')
 @click.option('--with-c-driver',
@@ -127,7 +127,7 @@ def release(jira_creds_file,
     Perform the steps associated with the release.
     """
 
-    check_libmongoc_version()
+    check_mongoc_version()
 
     # Read Jira credentials and GitHub token first, to check that
     # user has proper credentials before embarking on lengthy builds.
@@ -233,43 +233,43 @@ def release(jira_creds_file,
                                     release_notes_text, output_file, quiet)
 
 
-def check_libmongoc_version():
-    got_LIBMONGOC_REQUIRED_VERSION = None
-    got_LIBMONGOC_DOWNLOAD_VERSION = None
+def check_mongoc_version():
+    got_MONGOC_REQUIRED_VERSION = None
+    got_MONGOC_DOWNLOAD_VERSION = None
     with open("CMakeLists.txt", "r") as cmakelists:
         for line in cmakelists:
             match = re.match(
-                r"set\(LIBMONGOC_REQUIRED_VERSION\s+(.*?)\)", line)
+                r"set\(MONGOC_REQUIRED_VERSION\s+(.*?)\)", line)
             if match:
                 if 'TODO' in line:
                     click.echo(
-                        'Found TODO on LIBMONGOC_REQUIRED_VERSION line in CMakeLists.txt: {}'.format(line))
+                        'Found TODO on MONGOC_REQUIRED_VERSION line in CMakeLists.txt: {}'.format(line))
                     sys.exit(1)
-                got_LIBMONGOC_REQUIRED_VERSION = match.group(1)
+                got_MONGOC_REQUIRED_VERSION = match.group(1)
                 continue
             match = re.match(
-                r"set\(LIBMONGOC_DOWNLOAD_VERSION\s+(.*?)\)", line)
+                r"set\(MONGOC_DOWNLOAD_VERSION\s+(.*?)\)", line)
             if match:
                 if 'TODO' in line:
                     click.echo(
-                        'Found TODO on LIBMONGOC_DOWNLOAD_VERSION line in CMakeLists.txt: {}'.format(line))
+                        'Found TODO on MONGOC_DOWNLOAD_VERSION line in CMakeLists.txt: {}'.format(line))
                     sys.exit(1)
-                got_LIBMONGOC_DOWNLOAD_VERSION = match.group(1)
+                got_MONGOC_DOWNLOAD_VERSION = match.group(1)
                 continue
-    assert got_LIBMONGOC_DOWNLOAD_VERSION
-    assert got_LIBMONGOC_REQUIRED_VERSION
-    libmongoc_version_pattern = r'[0-9]+\.[0-9]+\.[0-9]+'
-    if not re.match (libmongoc_version_pattern, got_LIBMONGOC_DOWNLOAD_VERSION):
-        click.echo("Expected LIBMONGOC_DOWNLOAD_VERSION to match: {}, got: {}".format(
-            libmongoc_version_pattern, got_LIBMONGOC_DOWNLOAD_VERSION))
+    assert got_MONGOC_DOWNLOAD_VERSION
+    assert got_MONGOC_REQUIRED_VERSION
+    mongoc_version_pattern = r'[0-9]+\.[0-9]+\.[0-9]+'
+    if not re.match (mongoc_version_pattern, got_MONGOC_DOWNLOAD_VERSION):
+        click.echo("Expected MONGOC_DOWNLOAD_VERSION to match: {}, got: {}".format(
+            mongoc_version_pattern, got_MONGOC_DOWNLOAD_VERSION))
         sys.exit(1)
-    if not re.match (libmongoc_version_pattern, got_LIBMONGOC_REQUIRED_VERSION):
-        click.echo("Expected LIBMONGOC_REQUIRED_VERSION to match: {}, got: {}".format(
-            libmongoc_version_pattern, got_LIBMONGOC_REQUIRED_VERSION))
+    if not re.match (mongoc_version_pattern, got_MONGOC_REQUIRED_VERSION):
+        click.echo("Expected MONGOC_REQUIRED_VERSION to match: {}, got: {}".format(
+            mongoc_version_pattern, got_MONGOC_REQUIRED_VERSION))
         sys.exit(1)
-    if got_LIBMONGOC_DOWNLOAD_VERSION != got_LIBMONGOC_REQUIRED_VERSION:
-        click.echo("Expected LIBMONGOC_DOWNLOAD_VERSION ({}) to match LIBMONGOC_REQUIRED_VERSION ({})".format(
-            got_LIBMONGOC_DOWNLOAD_VERSION, got_LIBMONGOC_REQUIRED_VERSION))
+    if got_MONGOC_DOWNLOAD_VERSION != got_MONGOC_REQUIRED_VERSION:
+        click.echo("Expected MONGOC_DOWNLOAD_VERSION ({}) to match MONGOC_REQUIRED_VERSION ({})".format(
+            got_MONGOC_DOWNLOAD_VERSION, got_MONGOC_REQUIRED_VERSION))
         sys.exit(1)
 
 
@@ -384,8 +384,8 @@ def ensure_c_driver(c_driver_build_ref, with_c_driver, quiet):
     """
 
     if with_c_driver:
-        bson_h = os.path.join(with_c_driver, 'include/libbson-1.0/bson/bson.h')
-        mongoc_h = os.path.join(with_c_driver, 'include/libmongoc-1.0/mongoc/mongoc.h')
+        bson_h = os.path.join(with_c_driver, 'include/bson2/bson/bson.h')
+        mongoc_h = os.path.join(with_c_driver, 'include/mongoc2/mongoc/mongoc.h')
         if os.path.exists(bson_h) and os.path.exists(mongoc_h):
             return with_c_driver
         if not quiet: